Debugging Coveraged error

+++ SeniorCapstoneProject/Simulator.xaml.cs
@@ -32,7 +32,7 @@ namespace SeniorCapstoneProject
         Timer time;
         Thread timer;
         RobotMovedObserver movedObserver;
-        float coverage;
+        double coverage;
         BatteryDeadObserver dead;
 
         public static bool IsRunning
@@ -225,7 +225,7 @@ namespace SeniorCapstoneProject
         {
             int mins = secs / 60;
             
-            coverage = ((float)room.Vacuum.PointsVisited.Count / room.PointsInRoom())*100;
+            coverage = ((double)room.Vacuum.PointsVisited.Count / room.PointsInRoom())*100;
             this.Dispatcher.Invoke(() => { this.lblTime.Content = String.Format("Elapsed time: {0}:{1:00}", mins, secs % 60); this.lblBattery.Content = String.Format("Battery: {0:0.00}%", RobotVacuum.Vacuum.Battery.Percent); this.lblCoverage.Content = String.Format("Coverage: {0:0.000}%", coverage); }); //Updates the timer label.
 
             //Add coverage of 100% here
@@ -238,7 +238,7 @@ namespace SeniorCapstoneProject
 
         private void BatteryDead()
         {
-            IStatistics stats = new Statistics(time.Secs, coverage, RobotVacuum.Vacuum.Algorithm.Algorithm);
+            IStatistics stats = new Statistics(time.Secs, (float)coverage, RobotVacuum.Vacuum.Algorithm.Algorithm);
 
             IsRunning = false;
             time.Enabled = false; //disables the timer.
@@ -280,7 +280,7 @@ namespace SeniorCapstoneProject
         private void TerminateSimulation(string reason)
         {
 
-            IStatistics stats = new Statistics(time.Secs, coverage, RobotVacuum.Vacuum.Algorithm.Algorithm);
+            IStatistics stats = new Statistics(time.Secs, (float)coverage, RobotVacuum.Vacuum.Algorithm.Algorithm);
 
             IsRunning = false;
             time.Enabled = false; //disables the timer.
